Move LinkIconButton to common dir
