+++ extensions/ql-vscode/src/view/common/LinkIconButton.tsx
