+++ extensions/ql-vscode/src/view/variant-analysis/QueryDetails.tsx
@@ -1,6 +1,6 @@
 import { styled } from "styled-components";
 import { ViewTitle } from "../common";
-import { LinkIconButton } from "./LinkIconButton";
+import { LinkIconButton } from "../common/LinkIconButton";
 
 export type QueryDetailsProps = {
   queryName: string;
